refactor: kakaoService를 인터페이스로 추상화해 클라 구현체와의 결합 느슨하게 처리
