+++ README.md
@@ -90,7 +90,7 @@
 - [X] 카카오 api로 토큰 얻기
 - [ ] step1 피드백 반영
   - [ ] 외부 API 커스텀 예외 처리 
-  - [ ] Client 추상화
+  - [X] Client 추상화
   - [ ] 테스트 코드 작성
 - [ ] 주문(옵션, 수량, 메시지를 body로) 시 해당 상품 옵션의 수량 차감
   - [ ] 수량이 0이였다면 주문 불가
